experiments/image-segmentation/requirements.txt
dill
matplotlib
numpy
opencv-python
pandas
pillow
scikit-learn
spflow
